# leetcode-clone/backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Quiz Portal API",
    "description": "API Documentation for the Quiz Portal",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "description": "User login with rate limiting (max 10 requests per 15 min)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "test@example.com"
                },
                "password": {
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "examples": {
              "application/json": {
                "role": "student"
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "404": {
            "description": "User not found"
          },
          "429": {
            "description": "Rate limit exceeded",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "example": "Too many login attempts. Try again later."
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "type": "integer",
                "description": "Max requests allowed"
              },
              "RateLimit-Remaining": {
                "type": "integer",
                "description": "Remaining requests"
              },
              "RateLimit-Reset": {
                "type": "integer",
                "description": "Time when limit resets"
              }
            }
          },
          "503": {
            "description": "Fallback mode (Redis unavailable)",
            "examples": {
              "application/json": {
                "warning": "Redis unavailable, using in-memory rate limiting"
              }
            }
          }
        }
      }
    },
    "/auth/request-access": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "provider": {
                  "example": "any"
                },
                "department": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/auth/verify": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}